--- SharePointFramework/ProjectWebParts/src/components/RiskMatrix/IMatrixCell.ts
@@ -1,10 +1,0 @@
-import { MatrixCellType } from './MatrixCell'
-
-export interface IMatrixCell {
-  cellValue?: string
-  cellType: MatrixCellType
-  className: string
-  style?: React.CSSProperties
-  consequence?: number
-  probability?: number
-}
