+++ SharePointFramework/ProjectWebParts/src/components/RiskMatrix/MatrixCell/index.tsx
@@ -1,8 +1,8 @@
-import * as React from 'react'
+import React, { FunctionComponent } from 'react'
 import { IMatrixCellProps } from './types'
 import styles from './MatrixCell.module.scss'
 
-export const MatrixCell = ({ style, children }: IMatrixCellProps) => {
+export const MatrixCell: FunctionComponent<IMatrixCellProps> = ({ style, children }: IMatrixCellProps) => {
   return (
     <td className={styles.matrixCell} style={style}>
       <div className={styles.container}>{children}</div>
@@ -14,7 +14,7 @@ export interface IMatrixHeaderCellProps extends React.HTMLProps<HTMLElement> {
   label: string
 }
 
-export const MatrixHeaderCell = (props: IMatrixHeaderCellProps) => {
+export const MatrixHeaderCell: FunctionComponent<IMatrixHeaderCellProps> = (props: IMatrixHeaderCellProps) => {
   return (
     <td className={`${styles.matrixCell} ${styles.headerCell}`}>
       <span>{props.label}</span>
