+++ SharePointFramework/ProjectWebParts/src/components/RiskMatrix/RiskElement/index.tsx
@@ -1,9 +1,9 @@
-import * as React from 'react'
+import React, { FunctionComponent } from 'react'
 import { IRiskElementProps } from './types'
 import styles from './RiskElement.module.scss'
 import { RiskElementCallout } from './RiskElementCallout'
 
-export const RiskElement = ({ style, model, calloutTemplate }: IRiskElementProps) => {
+export const RiskElement: FunctionComponent<IRiskElementProps> = ({ style, model, calloutTemplate }: IRiskElementProps) => {
   const [callout, setCallout] = React.useState(null)
 
   const getTooltip = () => {
