+++ SharePointFramework/ProjectWebParts/src/components/RiskMatrix/types.tsx
@@ -1,3 +1,15 @@
+import { MatrixCellType } from './MatrixCell'
+
+export interface IMatrixCell {
+  cellValue?: string
+  cellType: MatrixCellType
+  className: string
+  style?: React.CSSProperties
+  consequence?: number
+  probability?: number
+}
+
+
 export interface IRiskMatrixProps {
   items?: RiskElementModel[]
   width?: number | string
